Drawer Added
Drawer Navigation has been added

You can now route between screens.

Need to figure out how to get rid of the tittle banner for the screens
