+++ App.js
@@ -1,25 +1,21 @@
-
-
-import { StyleSheet, Text, View, SafeAreaView, ImageBackground, StatusBar, Dimensions } from 'react-native';
-
+import 'react-native-gesture-handler'
+import { StyleSheet, Text, View, SafeAreaView, ImageBackground, StatusBar } from 'react-native';
+import { createDrawerNavigator  } from '@react-navigation/drawer';
+import { NavigationContainer } from '@react-navigation/native';
 import HomeScreen from './Views/HomeScreen';
+import PhotoOfTheDayScreen from './Views/PodScreen';
 
-const image = {
-  uri: 'https://wallpapercave.com/wp/wp5111092.jpg'
-}
 
-var {width} = Dimensions.get('window');
+const Drawer = createDrawerNavigator();
 
 export default function App() {
   return (
-    <ImageBackground source={image} resizeMode="cover">
-      <StatusBar barStyle='light-content' />
-      <SafeAreaView className="h-full">
-        <View>
-          <HomeScreen />
-        </View>
-      </SafeAreaView>
-    </ImageBackground>
+    <NavigationContainer >
+      <Drawer.Navigator initialRouteName="Home">
+        <Drawer.Screen name="Home" component={HomeScreen} />
+        <Drawer.Screen name="Photo Of The Day" component={PhotoOfTheDayScreen} />
+      </Drawer.Navigator>
+    </NavigationContainer>
   );
 }
