+++ Views/PodScreen.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import {View, Text} from 'react-native'
+
+
+export default function PhotoOfTheDayScreen({navigation}){
+    return(
+        <View>
+            <Text>
+                Photo Of The Day Screen
+            </Text>
+        </View>
+    )
+}
